requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
distro==1.9.0
docstring-parser==0.17.0
exceptiongroup==1.3.0
fastapi==0.120.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jinja2==3.1.6
jiter==0.11.1
markupsafe==3.0.3
numpy==2.0.2
pandas==2.3.3
pydantic==2.12.3
pydantic-core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
structlog==25.5.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
